Widen chapter indices to u16 for long tables of contents (#33)
* Widen chapter indices to u16 for long tables of contents

Books with more than 255 chapters wrapped or saturated everywhere the
chapter index travelled as a u8: the Chapters overview could not select
past entry 255, the reducer clamped the firmware's uncapped
ChapterCursor/Loaded corrections, and a restored position past chapter
255 snapped back.

Chapter indices and counts are now u16 end to end — RenderRequest,
StorageCommand, LibraryEvent, ReaderState, the reader store's loaded
chapter and UI count, the UiShell/colophon renderers, and both
emulators. The persisted AppStateRecord and POS.BIN already stored u16
chapters, so no on-disk format changes. Adds a reducer test driving
selection across the 255/256 boundary.

Ported from upstream 8da2d42.

* Track the current chapter by section marks past 256 TOC entries

The u16 chapter widening left one truncated map behind: the resident
chapter_page array held 256 per-chapter start pages, so
current_chapter_for_page() could never return a chapter past 255. After
jumping to a later chapter, refresh_chapter_tracking() recomputed the
resident chapter through the truncated map, so the colophon, the
Chapters list centering, and the persisted StoreProgress chapter all
fell back into 0..=255 on the exact long-TOC books the widening targeted.

A chapter's start page is always its spine's section start page, so the
map is now keyed by section instead of chapter: chapter_start marks each
resident section with the first TOC entry that opens it (stored as
chapter + 1 so the boot value stays all-zero and SD_LIBRARY stays in
.bss), and resolution takes the greatest marked section start not beyond
the page. Bounded by MAX_BOOK_SECTIONS (320), the map grows 512 -> 640
resident bytes but covers a table of contents of any length; the TOC.BIN
loader streams every chapter record instead of dropping ones past 256.

The resolver lives in proto::cache as pure functions (fw has no host
tests), with a regression test driving a 322-chapter TOC over 320
sections to chapter 321 and a second covering split, shared, and
unresolvable spines.

* Invalidate chapter marks when the section table changes

chapter_start runs parallel to book_sections, but clear_book_index()
and set_book_index() replaced the section table without dropping the
marks, and the chapter-map token only tracks source identity and
layout. A same-book, same-layout cache rebuild that slots sections
differently could pair the new table with marks derived from the old
one, misnaming the chapter in the colophon, the Chapters centering,
and the persisted position.

Clearing the index now always drops the marks. Setting it drops them
only when the incoming table actually differs from the resident one:
the book index reloads unchanged on every section crossing, so an
unconditional reset there would re-read the whole TOC from SD on the
reading path's hottest crossing.

* Resolve no chapter while the section-table marks are dropped

current_chapter_for_page() trusted its callers to check
chapter_start_ready, but since the marks can now be dropped while a
book stays loaded (a rebuilt section table), pairing stale marks with
the current sections is a state the resolver itself must refuse.
Return chapter 0 until the marks are re-derived.

+++ fw/src/views.rs
@@ -191,7 +191,7 @@ fn fill_chapters<'a>(
         let count = sd_library
             .chapter_count_for_ui()
             .max(1)
-            .min(chapters.len() as u8) as usize;
+            .min(chapters.len() as u16) as usize;
         for item in chapters.iter_mut().take(count) {
             // Empty titles render as numbered chapters in the contents view.
             *item = UiTocItem {
